temp/requirements.txt
requests
pandas
python-dotenv
google-cloud-bigquery
google-cloud-logging
google-cloud-pubsub
google-cloud-storage
google-cloud-dataproc
functions-framework                 